DOCTYPE resolution
The doctype tag has to remain at the top of the code page because this tells the browser what language it's reading. This is essential. Technically, my code preview is not a browser page so it's only getting the HTML that i'm making my website with. Ignore the doctype tag in the code preview. What matters is that the rest of my code is in order and works.